LayoutTests/editing/pasteboard/insert-font-with-size-and-css.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div contenteditable style="font-size: 25px;">WebKit</div></div>
<div id="content"><font size="7" style="font-size: 25px;">hello </font><font size="7" color="blue" style="font-size: 15px;">world </font></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests inserting a font element with both font-size CSS property and size attribute.\n'
    + 'WebKit should not strip just font-size property. Two instances of "hello world" should look identical to each other.');

Markup.dump('container', 'Insertion point');

var content = document.getElementById('content');
Markup.dump(content, 'Content to insert');

document.querySelector('div[contenteditable]').focus();
document.execCommand('InsertHTML', null, content.innerHTML);

Markup.dump('container', 'After insertion');

</script>
</body>
</html>
