LayoutTests/editing/pasteboard/insert-font-with-size-and-css-expected.txt

This tests inserting a font element with both font-size CSS property and size attribute.
CyberKit should not strip just font-size property. Two instances of "hello world" should look identical to each other.

Insertion point:
| <div>
|   contenteditable=""
|   style="font-size: 25px;"
|   "CyberKit"

Content to insert:
| <font>
|   size="7"
|   style="font-size: 25px;"
|   "hello "
| <font>
|   color="blue"
|   size="7"
|   style="font-size: 15px;"
|   "world "

After insertion:
| <div>
|   contenteditable=""
|   style="font-size: 25px;"
|   "hello "
|   <font>
|     color="blue"
|     style="font-size: 15px;"
|     "world <#selection-caret>"
|   "CyberKit"
